+++ src/vetportal/Pets.java
@@ -54,29 +54,4 @@ public class Pets {
     public int getPetOwner() {
         return petOwner;
     }
-
-    //Validate date of birth
-    public static Boolean validateDOB(String petDOB) throws ParseException {
-        //format date to match field input
-        DateFormat formatDate = new SimpleDateFormat("MM/dd/yy");
-        //get today's date
-        Date todayDate = new Date();
-        //correctly format today's date
-        Date cTodayDate = formatDate.parse(formatDate.format(todayDate));
-        //correctly format user input
-        Date cPetDOB = formatDate.parse(petDOB);
-
-        //compare the current date to input
-        //if today's date is after the pet's dob (pet's dob in past)
-        if (cTodayDate.compareTo(cPetDOB) > 0) {
-            //return true (valid DOB)
-            System.out.println("true");
-            return true;
-        }
-        //otherwise, return false (invalid dob)
-        else {
-            System.out.println("false");
-            return false;
-        }
-    }
 } //end of Pets
